+++ plonk/src/proof_system/snark.rs
@@ -448,8 +448,7 @@ where
         max_degree: usize,
         rng: &mut R,
     ) -> Result<Self::UniversalSRS, Self::Error> {
-        use ark_ec::{scalar_mul::fixed_base::FixedBase, CurveGroup};
-        use ark_ff::PrimeField;
+        use ark_ec::{CurveGroup, ScalarMul};
         use ark_std::{end_timer, start_timer, UniformRand};
 
         let setup_time = start_timer!(|| format!("KZG10::Setup with degree {}", max_degree));
@@ -465,14 +464,12 @@ where
             cur *= &beta;
         }
 
-        let window_size = FixedBase::get_mul_window_size(max_degree + 1);
-
-        let scalar_bits = E::ScalarField::MODULUS_BIT_SIZE as usize;
         let g_time = start_timer!(|| "Generating powers of G");
-        // TODO: parallelization
-        let g_table = FixedBase::get_window_table(scalar_bits, window_size, g);
-        let powers_of_g =
-            FixedBase::msm::<E::G1>(scalar_bits, window_size, &g_table, &powers_of_beta);
+        let powers_of_g = g
+            .batch_mul(&powers_of_beta)
+            .into_iter()
+            .map(E::G1::from)
+            .collect::<Vec<_>>();
         end_timer!(g_time);
 
         let powers_of_g = E::G1::normalize_batch(&powers_of_g);
@@ -1274,47 +1271,47 @@ pub mod test {
         let q_ecc = &pk.selectors[2 * GATE_WIDTH + 4];
         let circuit_poly = q_c
             + &oracles.pub_inp_poly
-            + oracles.wire_polys[0].mul(q_lc[0])
-            + oracles.wire_polys[1].mul(q_lc[1])
-            + oracles.wire_polys[2].mul(q_lc[2])
-            + oracles.wire_polys[3].mul(q_lc[3])
-            + oracles.wire_polys[0]
+            + (&oracles.wire_polys[0]).mul(q_lc[0])
+            + (&oracles.wire_polys[1]).mul(q_lc[1])
+            + (&oracles.wire_polys[2]).mul(q_lc[2])
+            + (&oracles.wire_polys[3]).mul(q_lc[3])
+            + (&oracles.wire_polys[0])
                 .mul(&oracles.wire_polys[1])
                 .mul(q_mul[0])
-            + oracles.wire_polys[2]
+            + (&oracles.wire_polys[2])
                 .mul(&oracles.wire_polys[3])
                 .mul(q_mul[1])
-            + oracles.wire_polys[0]
+            + (&oracles.wire_polys[0])
                 .mul(&oracles.wire_polys[1])
                 .mul(&oracles.wire_polys[2])
                 .mul(&oracles.wire_polys[3])
                 .mul(&oracles.wire_polys[4])
                 .mul(q_ecc)
-            + oracles.wire_polys[0]
+            + (&oracles.wire_polys[0])
                 .mul(&oracles.wire_polys[0])
                 .mul(&oracles.wire_polys[0])
                 .mul(&oracles.wire_polys[0])
                 .mul(&oracles.wire_polys[0])
                 .mul(q_hash[0])
-            + oracles.wire_polys[1]
+            + (&oracles.wire_polys[1])
                 .mul(&oracles.wire_polys[1])
                 .mul(&oracles.wire_polys[1])
                 .mul(&oracles.wire_polys[1])
                 .mul(&oracles.wire_polys[1])
                 .mul(q_hash[1])
-            + oracles.wire_polys[2]
+            + (&oracles.wire_polys[2])
                 .mul(&oracles.wire_polys[2])
                 .mul(&oracles.wire_polys[2])
                 .mul(&oracles.wire_polys[2])
                 .mul(&oracles.wire_polys[2])
                 .mul(q_hash[2])
-            + oracles.wire_polys[3]
+            + (&oracles.wire_polys[3])
                 .mul(&oracles.wire_polys[3])
                 .mul(&oracles.wire_polys[3])
                 .mul(&oracles.wire_polys[3])
                 .mul(&oracles.wire_polys[3])
                 .mul(q_hash[3])
-            + oracles.wire_polys[4].mul(q_o).neg();
+            + (&oracles.wire_polys[4]).mul(q_o).neg();
 
         // check that the polynomial evaluates to zero on the vanishing set
         let domain = Radix2EvaluationDomain::<E::ScalarField>::new(pk.domain_size())